crawler/requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
beautifulsoup4==4.7.1
boto3==1.9.119
botocore==1.12.120
Click==7.0
Flask==1.0.2
google-cloud-storage==1.14.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
mysql-connector-python==8.0.15
numpy==1.16.3
pluggy==0.9.0
py==1.8.0
pytest==4.3.0
python-dotenv==0.10.1
redis==3.2.0
requests==2.21.0
scipy==1.2.1
scikit-learn==0.20.3
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
